changes file names; added comments
